Elasticsearch-to-MySQL-master/Elasticsearch-to-MySQL-master/ElasticsearchToMysql/requirements.txt
elasticsearch==6.8.2
geopy==2.4.0
numpy==1.26.1
pandas==2.1.1
python_dateutil==2.8.2
Requests==2.31.0
SQLAlchemy==2.0.22